API config (#24)
Create a config system that uses
[pydantic-settings](https://docs.pydantic.dev/2.10/api/pydantic_settings/)
to load settings from a `config.toml` file, environment variables, and
Docker secrets. Dev configs are copied into the containers, but the prod
config is the default.

I also changed to a sync+restart strategy for watching in Docker Compose
because Granian doesn't reload if the app fails to instantiate because a
synced file has an error.